combine both verseions of eb into a single module
